changed log levels, removed .psd images and clean-up for the pre-release code.
Removed the .psd files in the images folder. Added .gitignore. Updated strings.xml to update the text of the fields on Settings screen. Removed ModuleListFragment. Changed the log level to debug from info for all the logs that output the ongoing status when sysncing the module data. In EditDetailsAtivity, in loadContentTask, commented the code in the case INPUT_TYPE where the casting for the AutoCompleteTextView is causing the application to crash. Updated the manifest file not to support large screens.