refactored to get rid of redundant constants and fixed few bugs
refactored to get rid of redundant constants defined in ImportContactsUtility, and
improper use of static HashMap of contacts map to module fields. Fixed the
ClassCastException when INPUT_TYPE case for phone is cast for an AutoTextViewField.
This bug causes the progress dialog to hang indefinetely when trying to edit a Module